+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `CadenzaClient#destroy`
+
+### Fixed
+- To avoid errors with active subscriptions when the iframe is not visible, a target window (for postMessage communication) is now required only for _sending_ a message (and not for receiving messages). 
 
 ## 2.9.0 - 2024-06-04
 ### Added
@@ -15,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `invalidateCaches` option for `CadenzaClient#reload`
 - Allow communication with parent Cadenza window (Previously, communication was possible only with an embedded Cadenza iframe.)
 - `cadenza()` overload with a single options parameter
-- `CadenzaClient#closeMe` function
+- `CadenzaClient#closeMe`
 - Documentation on listening when a custom application is closed (unload and visibilitychange events)
 
 ### Deprecated